Add ProposedDevice class
